Fix compilation without -Wno-sign-conversion

+++ src/mpack/mpack-common.h
@@ -257,7 +257,7 @@ MPACK_INLINE mpack_tag_t mpack_tag_array(int32_t count) {
     mpack_tag_t ret;
     mpack_memset(&ret, 0, sizeof(ret));
     ret.type = mpack_type_array;
-    ret.v.n = count;
+    ret.v.n = (uint32_t)count;
     return ret;
 }
 
@@ -266,7 +266,7 @@ MPACK_INLINE mpack_tag_t mpack_tag_map(int32_t count) {
     mpack_tag_t ret;
     mpack_memset(&ret, 0, sizeof(ret));
     ret.type = mpack_type_map;
-    ret.v.n = count;
+    ret.v.n = (uint32_t)count;
     return ret;
 }
 
@@ -275,7 +275,7 @@ MPACK_INLINE mpack_tag_t mpack_tag_str(int32_t length) {
     mpack_tag_t ret;
     mpack_memset(&ret, 0, sizeof(ret));
     ret.type = mpack_type_str;
-    ret.v.l = length;
+    ret.v.l = (uint32_t)length;
     return ret;
 }
 
@@ -284,7 +284,7 @@ MPACK_INLINE mpack_tag_t mpack_tag_bin(int32_t length) {
     mpack_tag_t ret;
     mpack_memset(&ret, 0, sizeof(ret));
     ret.type = mpack_type_bin;
-    ret.v.l = length;
+    ret.v.l = (uint32_t)length;
     return ret;
 }
 
@@ -294,7 +294,7 @@ MPACK_INLINE mpack_tag_t mpack_tag_ext(int8_t exttype, int32_t length) {
     mpack_memset(&ret, 0, sizeof(ret));
     ret.type = mpack_type_ext;
     ret.exttype = exttype;
-    ret.v.l = length;
+    ret.v.l = (uint32_t)length;
     return ret;
 }
